only remove the debug folder if targeting windows

+++ ports/vulkantools/portfile.cmake
@@ -29,7 +29,9 @@ vcpkg_install_copyright(FILE_LIST "${SOURCE_PATH}/LICENSE.txt")
 
 vcpkg_copy_tools(TOOL_NAMES vkvia vkconfig AUTO_CLEAN )
 
-file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug")
+if(VCPKG_TARGET_IS_WINDOWS)
+  file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug")
+endif()
 
 set(VCPKG_POLICY_EMPTY_INCLUDE_FOLDER enabled)
